+++ build.gradle
@@ -270,7 +270,8 @@ pitest {
   excludedClasses = [
     'uk.gov.hmcts.reform.cwrdapi.config.*',
     'uk.gov.hmcts.reform.cwrdapi.controllers.*',
-    'uk.gov.hmcts.reform.cwrdapi.service.impl.*',
+    'uk.gov.hmcts.reform.cwrdapi.service.impl.CaseWorkerDeleteServiceImpl',
+//    'uk.gov.hmcts.reform.cwrdapi.service.impl.ValidationServiceFacadeImpl',
     'uk.gov.hmcts.reform.cwrdapi.CaseWorkerRefApiApplication',
     'uk.gov.hmcts.reform.cwrdapi.servicebus.TopicPublisher'
   ]
